feat: add actions to card details
